fix '$\/]' in HTML::Entities::encode_entities
encode_entities() generates and evals custom code at runtime (as a
performance optimization). However, its code generation was too naive
and certain characters could be used to break out of the character class
regex:

 - `$` (dollar sign) would trigger perl's variable interpolation
 - `]` and `/` (the character class and regex delimiters, respectively)

The latter two were usually escaped, but not if they were preceded by
`\` in the input, even if that `\` was itself escaped by another `\`
(NB: this is why it is generally a mistake to handle escaping logic with
look-behinds: you need arbitrary-width look-behind to figure out whether
the current chain of backslashes is even or odd in length in order to
know what is being escaped or not).

The latter issue was fixed by doing a single pass over the input string
with no look-behind; the former by switching the delimiter to `'` (which
inhibits interpolation).

Fixes #44.